feat(wallet): add WebAssembly CLSAG transaction signer for browser wallet
Extract the node's self-contained CLSAG transaction implementation from
botho/src/transaction.rs into a standalone, wasm-compatible crate
(bth-transaction-clsag) and add a wasm-bindgen signer crate
(web/packages/wasm-signer) that compiles the build + CLSAG-sign path to
wasm32-unknown-unknown. The signer produces the exact bincode wire format the
node accepts via tx_submit, and the produced transaction round-trips through
the same Rust verifier the node runs.

botho/src/transaction.rs now re-exports bth-transaction-clsag unchanged, so all
existing crate::transaction::* / botho::transaction::* paths and the on-the-wire
format are preserved. A test-utils feature gates the test-only stub constructor
so botho's tests still build across the new crate boundary.

The spend/view private keys are passed into wasm by the caller and used only to
recover the one-time key and sign; they never leave the client.
